Refactor WAL read/write
Extract KVRecord/KVWriteRecord structs that we can reuse later for
SSTable serialisation (presuming we use a simple structure for SSTable
of KVs written one after the other).

Make read/write of WAL records part of the implementation methods of
the WALRecord struct. This doesn't add much apart from centralising
the logic of serialising and deserialising WAL records into the
struct.

Overall this is rather more complicated, but it will hopefully help
later.