Implement delete
This implements delete by adding KVValue, an enum that can represent
either a live value (KVValue::Some) or a deleted value
(KVValue::Deleted).

The memtable is updated to store KVValues so it is able to represent
deleted values.

KVRecord is updated to have a KVValue instead of a raw Vec<u8> `value`.
The KVRecord serialisation is changed to store a flag for deleted. The
KV_VERSION is also bumped, although we don't maintain any code for
data migration. Because this is a toy!

WAL and sstables are updated to support KVRecords with the new KVValue.
get/set are updated to do the right thing with KVValues and a new
deleted method is added to the top level API.

I chose to have all internal methods understand KVValue rather than
relying on deleted flags because it felt more likely that I'd get
the code right that way. It also allows for adding other flags to
the record more easily if required (eg, perhaps, version numbers if
we do MVCC).