Use generics with Into<Vec<u8>> for API ergonomics
This does result in extra copying but makes the use of the API a bunch
easier, as shown by the reduction in to_string, as_bytes, clone and
such in the tests.

As noted in the issue, we may be able to reduce copying using Cow,
but this is a decent improvement for now (and a bunch of the time we
copy the passed in values anyway, eg, so the memtable can own one).

https://github.com/mikerhodes/toykv/issues/1