wip scan - generic version for memtables and sstables
The code is a bit simpler if we use generics. The idea is that we will
have two fields, a Vec of memtables and sstables as the two iterators.
The MergeIterator can "know" that the memtables are "first" and iterate
over the two of Vecs. It's a bit more complex to iterate over the
two types, at least it will be once we write it, but for now it seems
a bit nicer.