Use dynamic dispatch over generics in MergeIterator; add memtable scan
It was quite awkward getting the generics setup in MergeIterator for
the scenario where a single MergeIterator only had a single type of
iterators inside it (sstables or Vecs during testing). The patch I
had in progress, from 5 months ago, never compiled and looked like
quite a mess when I came back to it.

Overall, I'd avoided Boxing and dyn for performance purposes, but I
think that (a) this is a toy thing and (b) there's a lot of disk access
going on so probably adding dynamic dispatch for easier to write (and
read) code is the right call.

This code is clear, but super-inefficient: the MergeIterator has to
take ownership over the iterators it is using because the iterator
could outlive ToyKV, as I've not figured lifetimes on all the relevant
bits. My hope is that now I have the code working, I can figure how
to do that. Because right now we copy the whole memtable just to scan
it! (The sstables are not so bad; we create new file handles instead).