Implement Block::decode(data) -> Block
Add tests written by Claude.

Ignore one test that is testing for things where we should have errors.
The Ignore line should help remind me to fix the edge cases that should
be erroring --- empty keys and empty values (the latter of which should
only come to us as the Deleted enum value).