perf: decrease data copying in create_and_seek_to_key
We manually run the iteration through the block we are using, such
that we don't need to create a second BlockIterator to return to
the caller.

From this, I can see that the optimal way of doing this is to have
a separate method to return the key as a slice of the block's data
directly, but that's more change than I'm up for right now.