Use an explicit sstable index file to track sstables
Add a JSON "index" file for the sstables we have in the database. This
is overall a lot simpler than relying on a file naming convention as
it is explicit about what files are included.

This allows us to have a much simpler file naming convention, using
just random strings.

It also lays the ground work for compaction where we will be doing more
extensive manipulation of files:

- The index maintains the current "state" of the sstables layers. Any
  in progress operations are "made visible" by writing their changes
  to this file. This means we can be messing with the files in the
  directory without confusing the data read path.
- Trash can be collected after compaction separately, by removing files
  that are no longer in the index.
- Broadly this extra flexibility is needed for adding extra levels and
  carrying out compaction more generally where we may have several
  temporary files in play at once while we build up a new layer, for
  example.