feat: sstable bloom filters
Issue: https://github.com/mikerhodes/toykv/issues/8

This commit adds a bloom filter to each sstable file.

When writing the sstable using TableBuilder, we add the keys to the
bloomfilter as we go. When we finalise the file, we write the bloom
filter at the end of the file, after the block metadata index.

TableIterator is able to read the bloom filter, and has new methods
might_contain_key and might_contain_hashed_key. The methods can be
used to determine whether a file might contain a key, or definitely
does not (per bloom filter behaviour).

The bloom filter is consulted during get single k/v operations in the
database. Instead of checking every sstable on disk for the key, the
bloom filter is used to efficiently filter out sstable files that
don't contain the key while iterating the TableIterators.

Bloom filters have no effect on scans.

Tests are not added for checking that bloom filters are used. Instead,
dbg!() calls are added (and commented out) that allow for checking that
the tables resulted in fewer on disk tables being consulted during
gets. This could be fixed by adding metrics for table iterations and
using these to confirm whether tables were filtered out.