Create SSTablesReader for each get call
There is no need to maintain an SSTablesReader when we are not actively
making a call to SSTable's `get`, as we clone the reader during `get`
anyway because each call to `get` needs its own file handles to use
when seeking.

This change makes it more explicit that we are re-creating the
SSTablesReader for each `get` call, it's no less efficient than what
we were doing already.

To optimise, we could add caches for bloom filter and metadata index
blocks for the TableIterator to use, to avoid the cost of reading these
on each `get` (or scan).