Create TableReader, a thread-safe reader for an sstable file
Issue: https://github.com/mikerhodes/toykv/issues/11

The idea with this is that TableIterator can maintain the state of
its iteration in a thread-unsafe manner, while TableReader can be
a thread-safe reader for an SSTable, using a Mutex around its
BufReader. The Mutex is used to ensure we do a complete read in one
go --- essentially that we don't break a read by interleaving seeks
from different threads.

The idea is that the TableReader can be created once and stored in
a cache to share between threads (inside an Arc).

This means that we can avoid re-opening (and reading the block
meta index and bloom filter) every time we carry out an operation.
I'd previously altered the code to do this, to be thread-safe
in a relatively simple way, but performance 4-5x worse (in the
examples/bench.rs benchmark, for 100000 writes by 1000 WAL threshold we
went from averaging 0.064ms/read to 2.5/read).

This TableReader + Mutex approach could be quite racy, but for now I
am avoiding the complexity of a pool of readers per file. I think we
can get a long way with this approach + a block cache.