Add thread-safety to both scan and write
Issue: https://github.com/mikerhodes/toykv/issues/11

First I had to switch the memtable to using the crossbeam skipmap
because it was thread-safe already. There didn't seem much point to
figuring out a concurrent map myself, so I used the one Mini-LSM used.

Then I had to extract out the "state" of the database --- the memtable,
WAL and current set of sstables. This needed protecting with a
read/write lock.

This took a long time to get right. Adding the read write lock to hold
the state was relatively easy, so I got the compiler happy with writing
from multiple threads quite quickly. It was `scan` that caused problems,
because the iterator that I needed to give to MergeIterator only
borrowed the cloned Arc for the memtable.

I worked out that I needed to make an object that could store the
memtable cloned Arc and hold the iterator alongside it --- so they
_would_ survive together, and the borrow should be fine. But for the
life of me I couldn't see how to tell the compiler that's what I was
trying to do, so in its view, the borrow still potentially outlived
the memtable clone.

It took a long while to track down --- in the end, after a few hours,
from a half remembered early part of Mini-LSM[1] --- that an extra
crate was required to create a self-referential struct where one field
could borrow another such that the semantics I needed were available.

I'm kind of surprised that plain rust couldn't do this, which I think
exposes that I still don't quite understand the reasons for needing
the extra crate with its unsafe code.

[1]: https://skyzh.github.io/mini-lsm/week1-02-merge-iterator.html#task-1-memtable-iterator