Move memtable+wal to module; remove KVWriteValue
Issue: https://github.com/mikerhodes/toykv/issues/16

Move memtable and WAL into a single struct, and move to the
src/memtable.rs module.

As part of this work, it was neccessary to change a couple of
KVWriteValue into KVValues. Originally, KVWriteValue was designed
so that we didn't have to clone the byte arrays for the values, it
just contained a slice unlike KVValue that has a Vec<u8>.

However, once I'd changed a couple of the KVWriteValue values to
KVValue, it became clearer that it was just as efficient to use
KVValues everywhere. The memtable takes a KVValue as its value, and
needs ownership. By passing the KVValue by reference into the WAL
write methods, it's just as efficient as the way KVWriteValue used
a slice inside the struct.

The change to KVValue changed a lot of tests that passed in slices,
but I think overall it makes the code better to remove a type that
had become unneeded.