Create new WAL rather than resetting it
Issue: https://github.com/mikerhodes/toykv/issues/16

This should help to prepare us for multiple memtables by removing
the assumption that we "reset" the memtable. However, we're not there
yet as the filename for memtables' WAL is hard-coded. So we need to
next work out how to record the WAL filenames somewhere for each
memtable (the ordering is important!).