Remove hardcoded WAL filename by adding WAL index file
This allows us to have different filenames for the WAL because we can
store the WAL filename in the WAL index file across restarts.

This sets us up for having a second memtable because that second
memtable (have a frozen and active memtable) can have a different
filename to the first.