Add v2 MergeIterator that uses a min-heap approach
Add a v2 MergeIterator that uses a min-heap rather than linear scan of
peekable iterators. This is a more standard merging algorithm, so far
as I can tell, for database merges like this.

`bench_scan` shows a reduction of about 10% per scan, from 8.5ms to
7.5ms.

Copy over the tests from v1 to ensure works the same. Inasmuch as
the tests test, obviously.

For now we retain the v1, just in case.