Implement frozen memtable, fail writes if need flush
This implements a second memtable, the frozen_memtable. When the
active memtable is full, it it moved to the frozen slot and a new
memtable is created for the active slot.

Instead of automatically flushing the frozen memtable, `set` and
`delete` will return an error if the active memtable gets full when
there is still an unflushed frozen memtable. This allows the user
to make use of backpressure --- if their application is writing too
fast they can reject writes so the storage layer can catch up.

The flush_oldest_memtable() method is added to toykv, allowing the
user to flush the frozen memtable, either periodically or when a
write returns a new error, ToyKVError::NeedFlush.

flush_oldest_memtable() currently takes out a write lock during the
whole flushing process, to make it simpler to implement --- no one
can come in and flush the table concurrently, so it's easier to get
right. This could be improved later if need be.

The main change I want to make is to make the number of frozen
memtables configurable. If this was combined with implementing a
max memtable size, this allows the user to tune the amount of memory
used for frozen memtables --- in particular, to allow "spikes" of
memtables when writes come in quickly.