Implement compaction v1
https://github.com/mikerhodes/toykv/issues/24

From the issue:

v1: compact all the L0 sstables into a single sstable. Do this in a
write lock for simplicity and leave the new sstable in the L0 list. So
the new memtables are put in the same place.

Checked that bench and bench_scan still perform the same.

One note is that in order to keep the performance the same, I had to
hardcode the bloomfilter expected keys to 1000 in the compact code.
Not quite sure how to generate the expected keys from the files being
compacted right now. PebbleDB has a bloomfilter that automatically
expands as items are added, but the rust implementations I can find
all expect you to know in advance. Perhaps I'm misreading Pebble's
code.