Implement compaction v2
Issue: https://github.com/mikerhodes/toykv/issues/24

This implements v2 of compaction from the above task:

> v2: Same thing, but don't do the whole compaction in a
> write lock. Take a read lock, read the list of files in
> L0. Compact them into the new sstable. Return the filename
> from the compact call. Take the write lock and remove
> the original files from the L0 list (they should be the
> tail --- perhaps check this for correctness then remove),
> then put the new file at the tail (so any newly written
> memtables->sstables are ahead of it, so keeping the order
> correct.

Use the same pattern as for writing out a memtable: the SSTables struct
has a "build compactor" method that can return a compaction task. The
build needs to happen with a read lock on the ToyKV state.

The compaction task can then be executed outside the state locks.

Finally, the compaction task returns a result that the SSTables struct
can use to commit the compaction to the store.

This commit also creates a CompactionPolicy trait and a
SimpleCompactionPolicy implementation, that encodes the policy to
compact all existing sstables into one big one.

Overall, this adds a few abstractions around the basic compact task,
but feels like it cleaves them correctly for future work.

This commit uses a new pattern, for me, the CompactionGuard which uses
the strict scope semantics of Rust to ensure we "release" the
state.compacting flag, even if the method early exits with an error.

Finally, a compacting field is added to state, to allow ToyKV to
reject concurrent compactions.