compaction v3: use CompactionPlan as input/output of compaction
Issue: https://github.com/mikerhodes/toykv/issues/24

We add the CompactionPlan struct which mirrors the levels in the
sstable index. We use it for inputs to the compaction task and the
output in the result.

This allows us to have the input being both l0 and l1 tables, as well
as the output being either l0 or l1 tables. For now, the code just
uses l0.

The next step is compacting to l1, and then updating scan and get to
use the l1 (single) table.