compaction v3: compact to l1; add hacky get/scan via SSTablesReader
Issue: https://github.com/mikerhodes/toykv/issues/24

Here we compact to l1, and implement a hacky get/scan by using the
knowledge that we are compacting to a single l1 table to just put
the l1 table at the end of the existing l0 table readers vec.

This isn't a good move long term, but it works to allow tests to
pass for now. Really we need to handle l1 and above differently to l0
as they are different things --- l1 being a single multi-file sorted
run, whereas l0 is multiple single-file sorted runs.