compaction v3: add TableReader first/last key
Issue: https://github.com/mikerhodes/toykv/issues/24

Move TableReader into src/table/reader.rs to make it easier to
navigate to and also so that we can add specific tests without
making any one file too large.

Adds key_range function to TableReader so that ConcatIterator can
determine the key range of each sstable it's using. The first/last
keys are pulled from the blockmeta when the TableReader struct
is created, so it's efficient to read these whenever we need to.