compaction v3: concatiterator scan to key to start iteration
Issue: https://github.com/mikerhodes/toykv/issues/24

Start writing ConcatIterator. Write ConcatIterator::new() which takes
a sorted run of TableReaders, and a lower and upper bound. It returns
a ConcatIterator that is ready to start iterating --- it's got a
TableIterator wrapped around the right TableReader for the lower
bound.

For now the right sstable is found via a linear scan. This will suffice
for smaller databases (eg, at 256MB per table, a 10GB database is
still only 40 sstables). But we should probably make this better
later.

Add tests for the method that finds the right TableReader in the
sorted run. Make TableBuilder pub(crate) so we can use it in the tests.

Add a TableIterator::new_bounded_with_tablereader method as most
other range methods support bounds. Albeit this one doesn't support
Excluded bounds as yet (need to fix up the binary search that
finds the right block) --- so this is just a refactor to add support
for using Bounds as inputs, it doesn't support any new features (ie,
Bounds::Excluded).