compaction v3: Allow embedding MergeIterators for scanning
Issue: https://github.com/mikerhodes/toykv/issues/24

This commit alters scanning to use a single MergeIterator returned
from the SSTables::iter method during scanning. That returned
MergeIterator for the sstables is added to the MergeIterator that
ToyKV::scan returns, which still results in the correct results for
scan because they were designed to make them composable from the
start.