compact v3: Use Bounds for TableIterator
Issue:

Use TableIterator::new_bounded_with_tablereader everywhere outside
impl TableIterator itself, although still with the caveat that we
don't support Bound::Excluded.

This allows simplifying a few places where we matched on the Option for
the key and called the relevant new method for TableIterator to instead
use match to create the Bound from the key Option. This makes the
code creating the iterators themselves much simpler, which decreases
the complexity of the methods that create them.