compaction v4: drop deleted kv pairs during compaction
Issue: https://github.com/mikerhodes/toykv/issues/24

This tweaks the compaction writing process to avoid adding deleted
items to the new tables. The actual code change for that is tiny.

Instead, the largest changes are to the tests. We move compaction.rs
to src/compaction/mod.rs and move tests/compaction_v4_test.rs to
src/compaction/compaction_v4_test.rs so we can access the lower-level
types for writing tests. For testing the deleted kv pairs this is
required, because without it, we can't differentiate between the
desired state --- we dropped the kv pair completely --- and we carried
over the kv pair with its KValue::Deleted value. The ToyKV interface
returns the same regardless.

So we need to work at the SSTables layer instead, as this allows us
to differentiate that key case. It also gives us much more control
over the sstables themselves, because we don't have ToyKV managing
active/frozen memtables for us --- instead we make our own memtables
and explicitly flush them. This gives us much greater control.

For now, this is only used for the delete test where it's strictly
required, but later I should port the approach to the other tests
because it's much more explicit about what files we have and what
we expect.

Even so, the other tests are updated to use the SSTablesIndex class
to read the index from disk to check that the live tables are correct
in the index JSON, rather than merely counting the files on disk.

Definitely better tested all round, as well as the function change of
dropping deleted items.