test: add test for failing block checksum
Issue: https://github.com/mikerhodes/toykv/issues/14

I could see that checksum failure was causing the panic it should be
while I was fixing broken tests after adding the xxhash checksum to
blocks. However, I forgot to add an explicit test for it. So here is
a regression test for that functionality.

Also fixup the README to note we have checksums now.